requirements.txt
streamlit
numpy
matplotlib
pillow
scikit-image
scipy